requirements.txt
alpaca-trade-api==0.46
ta
sklearn
python-dotenv==0.13.0